添加了很简单的资源文件上传功能 （md5验证，资源分类）

+++ src/main/java/com/hyp/myweixin/pojo/modal/WeixinResource.java
@@ -51,7 +51,7 @@ public class WeixinResource {
     /**
      * 文件大小
      */
-    private Double size;
+    private Long size;
 
     /**
      * 文件MD5值
